Reworked events endpoint to a single shared endpoint; WiP front end view of picture duplicates; Remove active picture from redux state

+++ reactive-albums-api/src/main/kotlin/nl/juraji/reactive/albums/query/projections/DuplicateMatchProjection.kt
@@ -8,5 +8,5 @@ data class DuplicateMatchProjection(
         @Id val id: String,
         val pictureId: String,
         val targetId: String,
-        val similarity: Float,
+        val similarity: Int,
 )
