+++ reactive-albums-api/src/main/resources/db/migration/projections/V004__add_duplicate_match_projection_table.sql
@@ -1,8 +1,8 @@
 create table DuplicateMatchProjection
 (
-    id         varchar(64)  not null,
+    id         varchar(64) not null,
     pictureId  varchar(64) not null,
     targetId   varchar(64) not null,
-    similarity float        not null,
+    similarity int         not null,
     primary key (id)
 )
